pyproject.toml
[project]
name = "skill-fleet"
version = "0.3.0"
description = "Hierarchical skills taxonomy + DSPy workflow prototype"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Qredence" }]
license = { text = "Apache-2.0" }
keywords = ["agents", "skills", "dspy", "fastapi", "cli", "workflow"]
dependencies = [
  "asyncpg>=0.31.0",
  "aiosqlite>=0.17.0",
  "click>=8.3.1",
  "datasets>=4.4.2",
  "dspy>=3.1.2,<4",
  "fastapi[standard]>=0.128.0",
  "google-genai>=1.56.0",
  "httpx>=0.28.1",
  "litellm[proxy]>=1.80.16",
  "mlflow>=3.8.1",
  "openai>=2.14.0",
  "psycopg[binary]>=3.2.5",
  "pydantic>=2.12.5",
  "pyyaml>=6.0.3",
  "python-dotenv>=1.2.1",
  "rich>=13.7.1",
  "sqlalchemy>=2.0.45",
  "tomli>=2.0",
  "typer>=0.21.1",
  "typing-extensions>=4.0",
  "uvicorn[standard]>=0.31.1",
  "prompt-toolkit>=3.0.52",
  "ty>=0.0.12",
  "gepa>=0.0.24",
  "questionary>=2.0.0",
  "greenlet>=3.3.0",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
]

[project.urls]
Homepage = "https://github.com/Qredence/skill-fleet"
Repository = "https://github.com/Qredence/skill-fleet"
"Bug Tracker" = "https://github.com/Qredence/skill-fleet/issues"
Documentation = "https://github.com/Qredence/skill-fleet/tree/main/docs"

[project.scripts]
fleet-agent = "skill_fleet.cli:cli_entrypoint"
skill-fleet = "skill_fleet.cli:cli_entrypoint"

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1",
  "pytest>=8.0.0",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.0.0",
  "ruff>=0.8.0",
]

[dependency-groups]
dev = [
  "bandit[toml]>=1.8.0",
  "pre-commit>=4.0.1",
  "pytest>=8.0.0",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.0.0",
  "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skill_fleet"]
include = ["src/skill_fleet/config/**", "src/skill_fleet/_seed/**"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
  "integration: marks tests as integration tests (may require LLM access)",
  "slow: marks tests as slow running",
]
filterwarnings = [
  "ignore:Pydantic serializer warnings:UserWarning",
  # DeprecationWarning from aiohttp/connector.py in python 3.13
  "ignore:enable_cleanup_closed ignored:DeprecationWarning",
]

[tool.ruff]
src = ["src", "skills"]
target-version = "py312"
line-length = 100
exclude = ["skills/**"]

[tool.ruff.lint]
# Enable comprehensive rule set for 2026 best practices
select = [
  "E",     # pycodestyle errors
  "F",     # pyflakes
  "I",     # isort (import sorting)
  "UP",    # pyupgrade (modernize Python syntax)
  "B",     # flake8-bugbear (detect common bugs)
  "D",     # pydocstyle (docstring conventions)
  "N",     # pep8-naming (naming conventions)
  "C4",    # flake8-comprehensions (optimize comprehensions)
  "SIM",   # flake8-simplify (simplify code)
  "TC001", # Type checking - typing imports in runtime code
  "TC002", # Type checking - immutable default in mutable context
  "TC003", # Type checking - import from typing instead of collections.abc
  "D100",  # Missing module docstring
  "D101",  # Missing class docstring
  "D102",  # Missing method docstring
  "D103",  # Missing function docstring
  "D104",  # Missing package docstring
]
ignore = [
  "E501", # Line length handled by formatter
  "D212", # Alternative to D213 (multi-line docstring style)
  "D203", # Alternative to D211 (blank line before class)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D"] # Disable docstring checks for tests
# B008: Do not perform function call in default arguments
# Standard pattern for CLI libraries (typer, click)
"src/skill_fleet/cli/*.py" = [
  "B008",
  "D413",
  "D401",
  "SIM102",
] # Allow flexible docstring formatting and nested ifs in CLI
"src/skill_fleet/cli/**/*.py" = [
  "D413",
  "D401",
  "TC002",
  "TC003",
  "SIM102",
] # Relax docstring, type-checking, and simplification rules in CLI submodules
"src/skill_fleet/cli/interactive_cli.py" = [
  "TC001",
] # Allow runtime imports for type hints
"src/skill_fleet/cli/commands/db_sync.py" = [
  "D400",
  "D415",
] # Relax docstring ending rules for module docstrings
"src/skill_fleet/cli/exceptions.py" = ["N818"] # Allow CLIExit exception name

# API and core modules - relax type-checking rules for runtime imports
"src/skill_fleet/api/*.py" = [
  "TC001",
  "TC002",
  "TC003",
  "SIM105",
  "D401",
  "D107",
  "N818",
  "SIM110",
]
"src/skill_fleet/api/**/*.py" = ["TC001", "TC002", "TC003", "D417", "SIM110"]
"src/skill_fleet/api/exceptions.py" = ["N818"] # Allow APIException naming
"src/skill_fleet/analytics/*.py" = ["D107", "N803", "N806"]
"src/skill_fleet/core/**/*.py" = [
  "TC003",
  "SIM105",
  "D401",
  "D107",
  "D301",
  "SIM108",
]
"src/skill_fleet/common/*.py" = ["TC003", "D107", "D301"]
"src/skill_fleet/taxonomy/*.py" = [
  "SIM105",
  "SIM108",
  "SIM110",
  "SIM222",
  "D301",
]
"src/skill_fleet/validators/*.py" = ["D107", "D301"]

# Scripts - relaxed rules for utility scripts
"scripts/**/*.py" = [
  "D",
  "TC001",
  "TC002",
  "TC003",
  "SIM105",
  "SIM108",
  "SIM110",
  "SIM222",
  "D401",
  "E402",
]

# Tests - relaxed rules
"tests/**/*.py" = ["D", "TC001", "TC003", "SIM114", "SIM117", "SIM222", "N806"]

[tool.ruff.format]
# Explicit formatting configuration
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false

[tool.bandit]
# Bandit security linter configuration
exclude_dirs = ["tests", ".venv", "venv", ".git"]
skips = [
  # Skip assert_used - we use assertions for internal consistency checks
  "B101",
  # Skip hardcoded_password_func - we use environment variables for secrets
  "B105",
  "B106",
]
assert_used = false

[tool.ty.environment]
# Python version to assume (affects syntax and stdlib types)
python-version = "3.12"

# Path to Python environment for third-party modules (auto-detected by ty)
# python = "./.venv"

# Target platform ("win32", "darwin", "linux", "android", "ios", "all")
# ty will auto-detect from current platform
# python-platform = "linux"

# First-party module search paths (priority order)
root = ["./src"]

[tool.ty.rules]
# Set rule severities: "error", "warn", or "ignore"

# Critical errors
division-by-zero = "error"
possibly-unresolved-reference = "error"
invalid-argument-type = "error"
unsupported-operator = "error"

# Code quality warnings
unused-ignore-comment = "warn"
redundant-cast = "warn"

# Relaxed rules for false positives
index-out-of-bounds = "warn"

[tool.ty.src]
# Include specific files/directories
include = ["src", "tests"]

[tool.ty.terminal]
output-format = "full"
error-on-warning = false

# Relaxed rules for tests
[[tool.ty.overrides]]
include = ["tests/**"]

[tool.ty.overrides.rules]
possibly-unresolved-reference = "warn"
